requirements.txt
requests==2.31.0
cryptography==42.0.0
python-dotenv==1.0.0
numpy==1.26.0
pydantic==2.6.0
fake-useragent==1.4.0